specifying the type of relationship and checking

+++ backend/internal/models/socialUser/socialUser.go
@@ -5,7 +5,6 @@ import (
 	appErr "backend/internal/errors/appError"
 	"backend/internal/models/user"
 	"database/sql"
-	"fmt"
 )
 
 type SocialUser struct {
@@ -34,34 +33,30 @@ func AddFriend(userID, friendID uint64) error {
 		return appErr.BadRequest("user is banned")
 	}
 
-	isBlocked, err := checkBlock(userID, friendID)
+	db := pgDB.GetDB()
+	status, err := GetRelations(userID, friendID)
 	if err != nil {
 		return err
 	}
-	if isBlocked {
+	if status != nil && *status == "blocked by target" {
 		return appErr.BadRequest("you are blocked by this user")
-	}
-
-	db := pgDB.GetDB()
-	var status string
-	query := `
-		SELECT fs.name FROM friends f 
-        JOIN friend_statuses fs ON f.status_id = fs.id
-        WHERE 
-			(f.friend_1_id = $1 AND f.friend_2_id = $2) 
-            OR (f.friend_1_id = $2 AND f.friend_2_id = $1)
-	`
-	err = db.QueryRow(query, userID, friendID).Scan(&status)
-	if err == sql.ErrNoRows {
-		_, err = db.Exec(`INSERT INTO friends (friend_1_id, friend_2_id, status_id)
-		VALUES ($1, $2, (SELECT id FROM friend_statuses WHERE name = 'request'))`, userID, friendID)
-	} else if err == nil && status == "request" {
+	} else if status != nil && *status == "blocked" {
+		return appErr.BadRequest("you blocked this user")
+	} else if status != nil && *status == "accepted" {
+		return appErr.BadRequest("you are already friends")
+	} else if status != nil && *status == "outgoing request" {
+		return appErr.BadRequest("friend request has already been sent")
+	} else if status != nil && *status == "incoming request" {
 		_, err = db.Exec(`UPDATE friends SET status_id = (SELECT id FROM friend_statuses WHERE name = 'accepted')
 		WHERE (friend_1_id = $1 AND friend_2_id = $2) OR (friend_1_id = $2 AND friend_2_id = $1)`, userID, friendID)
+	} else {
+		_, err = db.Exec(`INSERT INTO friends (friend_1_id, friend_2_id, status_id)
+		VALUES ($1, $2, (SELECT id FROM friend_statuses WHERE name = 'request'))`, userID, friendID)
 	}
 	if err != nil {
 		return appErr.InternalServerError("internal server error")
 	}
+
 	return nil
 }
 
@@ -72,10 +67,16 @@ func RemoveFriend(userID, friendID uint64) error {
 		return err
 	}
 	db := pgDB.GetDB()
-	_, err = db.Exec(`DELETE FROM friends WHERE (friend_1_id = $1 AND friend_2_id = $2)
-	OR (friend_1_id = $2 AND friend_2_id = $1)`, userID, friendID)
+	status, err := GetRelations(userID, friendID)
 	if err != nil {
-		return appErr.InternalServerError("internal server error")
+		return err
+	}
+	if status != nil {
+		_, err = db.Exec(`DELETE FROM friends WHERE (friend_1_id = $1 AND friend_2_id = $2)
+		OR (friend_1_id = $2 AND friend_2_id = $1)`, userID, friendID)
+		if err != nil {
+			return appErr.InternalServerError("internal server error")
+		}
 	}
 	return nil
 }
@@ -87,10 +88,18 @@ func BlockUser(userID, targetID uint64) error {
 		return err
 	}
 	db := pgDB.GetDB()
-	_, err = db.Exec(`DELETE FROM friends WHERE (friend_1_id = $1 AND friend_2_id = $2)
-	OR (friend_1_id = $2 AND friend_2_id = $1)`, userID, targetID)
+	status, err := GetRelations(userID, targetID)
 	if err != nil {
-		return appErr.InternalServerError("inernal server error")
+		return err
+	}
+	if status != nil && *status == "blocked" {
+		return appErr.BadRequest("you have already blocked this user")
+	} else if status != nil && *status != "blocked by target" {
+		_, err = db.Exec(`DELETE FROM friends WHERE (friend_1_id = $1 AND friend_2_id = $2)
+		OR (friend_1_id = $2 AND friend_2_id = $1)`, userID, targetID)
+		if err != nil {
+			return appErr.InternalServerError("inernal server error")
+		}
 	}
 
 	_, err = db.Exec(`INSERT INTO friends (friend_1_id, friend_2_id, status_id)
@@ -108,32 +117,52 @@ func UnblockUser(userID, targetID uint64) error {
 		return err
 	}
 	db := pgDB.GetDB()
-	_, err = db.Exec(`DELETE FROM friends WHERE friend_1_id = $1 AND friend_2_id = $2 
-    AND status_id = (SELECT id FROM friend_statuses WHERE name = 'blocked')`, userID, targetID)
+	status, err := GetRelations(userID, targetID)
 	if err != nil {
-		fmt.Println(err)
-		return appErr.InternalServerError("inernal server error")
+		return err
 	}
+	if status != nil && *status == "blocked" {
+		_, err = db.Exec(`DELETE FROM friends WHERE friend_1_id = $1 AND friend_2_id = $2 
+        AND status_id = (SELECT id FROM friend_statuses WHERE name = 'blocked')`, userID, targetID)
+		if err != nil {
+			return appErr.InternalServerError("inernal server error")
+		}
+	}
+	
 	return nil
 }
 
-// checking whether the user is in the block list of the target
-func checkBlock(userID, targetID uint64) (bool, error) {
+// Get relations between two users
+func GetRelations(userID, targetID uint64) (*string, error) {
 	db := pgDB.GetDB()
-
-	var blockingStatus string
-	checkBlockingQuery := `
-		SELECT fs.name FROM friends f 
-		JOIN friend_statuses fs ON f.status_id = fs.id
-		WHERE f.friend_1_id = $1 AND f.friend_2_id = $2 AND fs.name = 'blocked'
+	query := `
+		SELECT 
+			CASE 
+				WHEN f.status_id = (SELECT id FROM friend_statuses WHERE name = 'accepted') THEN 'accepted'
+				WHEN f.status_id = (SELECT id FROM friend_statuses WHERE name = 'blocked') 
+					AND f.friend_1_id = $1 THEN 'blocked'
+				WHEN f.status_id = (SELECT id FROM friend_statuses WHERE name = 'blocked') 
+					AND f.friend_2_id = $1 THEN 'blocked by target'
+				WHEN f.status_id = (SELECT id FROM friend_statuses WHERE name = 'request') 
+					AND f.friend_1_id = $1 THEN 'outgoing request'
+				WHEN f.status_id = (SELECT id FROM friend_statuses WHERE name = 'request') 
+					AND f.friend_2_id = $1 THEN 'incoming request'
+				ELSE NULL
+			END AS friendship_status
+		FROM friends f
+		WHERE 
+			(f.friend_1_id = $1 AND f.friend_2_id = $2)
+			OR (f.friend_1_id = $2 AND f.friend_2_id = $1)
+		LIMIT 1
 	`
-	err := db.QueryRow(checkBlockingQuery, targetID, userID).Scan(&blockingStatus)
-	if err != nil && err == sql.ErrNoRows {
-		return false, nil
-	} else if err == nil && blockingStatus == "blocked" {
-		return true, nil
+
+	var friendshipStatus string
+	err := db.QueryRow(query, userID, targetID).Scan(&friendshipStatus)
+	if err == sql.ErrNoRows {
+		return nil, nil
 	} else if err != nil {
-		return false, appErr.InternalServerError("internal server error")
+		return nil, appErr.InternalServerError("internal server error")
 	}
-	return false, nil
+
+	return &friendshipStatus, nil
 }
